relocate python scripts
